GGUF FE: move parameter-pruning passes to their own file, trim comments
Move SnapshotLiveParameters and PruneParametersOrphanedSince out of
translate_session.cpp into pass/prune_orphaned_parameters.{hpp,cpp},
matching the existing pass/lower_set_rows_stateless.{hpp,cpp} layout.
In their own named namespace (ov::frontend::gguf::pass), they can use
the normal OPENVINO_MODEL_PASS_RTTI macro instead of the hand-written
get_type_info_static()/get_type_info() the anonymous-namespace version
needed to dodge the -Werror=attributes issue.

Also trims several overly long comments added this session (in
translate_session.cpp and the new pass files) down to what's needed to
follow the code.

Verified: ov_gguf_frontend_tests passes 245/245; both new/changed files
still compile clean with -Werror added to their exact build commands.
test-recurrent-state-rollback still passes 5/5.
